# visual-compiler/backend/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.IDRequest:
    properties:
      project_name:
        description: User's project name
        type: string
      users_id:
        description: Represents the User's ID from frontend
        type: string
    required:
    - project_name
    - users_id
    type: object
...
  handlers.ProjectRequest:
    properties:
      pipeline:
        additionalProperties: {}
        description: User's saved pipeline
        type: object
      project_name:
        description: User's project name
        type: string
      users_id:
        description: Represents the User's ID from frontend
        type: string
    required:
    - pipeline
    - project_name
    - users_id
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Documentation for Visual Compiler API endpoints
  title: Visual Compiler API
  version: "1.0"
paths:
...
  /users/deleteProject:
    delete:
      consumes:
      - application/json
      description: Deletes a project and all associated fields in every collection
        in the database
      parameters:
      - description: Delete Project
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.IDRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted project
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to delete project
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a project
      tags:
      - Users
  /users/getProject:
    get:
      consumes:
      - application/json
      description: Gets all the details about a user's project in every collection
        in the database
      parameters:
      - description: Name of project
        in: query
        name: project_name
        required: true
        type: string
      - description: User's ID
        in: query
        name: users_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Retrieved all project details
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input/Conversion failed
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Error in reading from specific collection
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a project
      tags:
      - Users
  /users/getProjects:
    get:
      consumes:
      - application/json
      description: Gets all the user's project names
      parameters:
      - description: User's ID
        in: query
        name: users_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Retrieved all project names
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all user's projects
      tags:
      - Users
...
  /users/save:
    post:
      consumes:
      - application/json
      description: Saves the project name in the Users table
      parameters:
      - description: Save User's Defined Project Name
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.IDRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Project added successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Project name already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Save project name
      tags:
      - Users
  /users/savePipeline:
    post:
      consumes:
      - application/json
      description: Saves the project's pipeline to be shown when the user opens that
        project on the frontend
      parameters:
      - description: Save User's Pipeline
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Pipeline saved
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Save project pipeline from frontend
      tags:
      - Users
...
swagger: "2.0"